move clock_time_to_tick to time.c
