+++ components/libc/compilers/common/sys/time.h
@@ -86,6 +86,7 @@ int settimeofday(const struct timeval *tv, const struct timezone *tz);
 int clock_getres  (clockid_t clockid, struct timespec *res);
 int clock_gettime (clockid_t clockid, struct timespec *tp);
 int clock_settime (clockid_t clockid, const struct timespec *tp);
+int clock_time_to_tick(const struct timespec *time);
 #endif /* RT_USING_POSIX */
 
 #ifdef __cplusplus
